pyproject.toml
[tool.coverage.run]
branch = true
source = ["aws_secrets"]
relative_files = true

[tool.coverage.report]
exclude_lines = []
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'coverage/html' --cov-report xml:'coverage/coverage.xml'"

[tool.poetry]
name = "aws-ssm-secrets-cli"
version = "2.4.1"
description = "AWS Secret CLI for manage SSM SecureString and SecretsManager"
license = "MIT"
authors = ["Lucas Vieira <lucas.vieira94@outlook.com>"]
maintainers = ["Lucas Vieira <lucas.vieira94@outlook.com>"]
readme = "README.md"
repository = "https://github.com/lucasvieirasilva/aws-ssm-secrets-cli"
keywords = ['AWS', 'SECRETS', 'SSM', 'MANAGER']
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
]

[[tool.poetry.packages]]
include = "aws_secrets"

[tool.poetry.scripts]
aws-secrets = "aws_secrets.cli.cli:cli"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
boto3 = "^1.3"
click = "^8.1.7"
jsonschema = "^3.2.0"
aws-encryption-sdk = "^3.0.0"
prompt-toolkit = "^3.0.47"
ruamel-yaml = "^0.18.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-cov = "^4.1.0"
pyfakefs = "^5.2.4"
flake8 = "^6.1.0"
autopep8 = "^2.0.4"
pytest-sugar = "^0.9.7"
toml = "^0.10.2"
python-semantic-release = "^8.0.8"
pre-commit = "^3.4.0"
commitizen = "^3.8.0"

[build-system]
requires = ["poetry-core==1.6.1"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
major_on_zero = false
branch = "master"
upload_to_release = true
build_command = "pip install poetry && poetry build"
commit_message = "chore(release): {version} [skip ci]"
tag_format = "v{version}"
